Add files via upload
The purpose of this data exercise was to use bank data on 5000 customers to predict the impact of whether a customer is an active user of online services and the impact of the customer having a bank-issued credit card on a customer’s response to a personal loan campaign. Specifically, we want to see if the bank customer is more likely to take out a loan if they have an online account with the bank and/or if they have a credit card issued by the bank.

Naive Bayes was used to determine conditional probabilities and the confusion matrix was utilized to determine the accuracy and predictive capacity of the model.